receiving commentViews with replies from CommentViewRepository
