+++ lib/modules/proposals/pages/proposal_comments/proposal_comments.page.dart
@@ -1,55 +1,19 @@
 import 'package:demos_app/config/routes/application.dart';
 import 'package:demos_app/config/routes/routes.dart';
-import 'package:demos_app/core/enums/space_role.enum.dart';
 import 'package:demos_app/modules/proposals/pages/proposal_comments/models/comment_view.model.dart';
+import 'package:demos_app/modules/proposals/pages/proposal_comments/repositories/comment_view.repository.dart';
 import 'package:demos_app/modules/proposals/pages/proposal_comments/widgets/member_comment.widget.dart';
-import 'package:demos_app/modules/spaces/models/member.view.dart';
+import 'package:demos_app/modules/proposals/pages/proposal_details/bloc/proposal_details.bloc.dart';
 import 'package:fluro/fluro.dart';
 import 'package:flutter/material.dart';
-
-Future<List<CommentView>> _getFakeComments() async {
-  final member = MemberView(
-      userId: '1',
-      role: SpaceRole.admin,
-      memberCreatedAt: '',
-      profilePictureKey:
-          'avatars/bc0e5b62-117d-45c4-af07-51818ccd3e0b.8846.jpg',
-      memberName: 'Gendo Ikari');
-
-  return <CommentView>[
-    CommentView(
-        commendId: '1',
-        member: member,
-        createdAt: '2022-2-15T02:54:29.907Z',
-        upVotesCount: 5,
-        downVotesCount: 2,
-        content: 'Estoy a favor porque ...',
-        repliesCount: 0),
-    CommentView(
-        commendId: '1',
-        member: member,
-        createdAt: '2022-2-15T02:54:29.907Z',
-        upVotesCount: 2,
-        downVotesCount: 0,
-        content: 'Bla bla bla...',
-        repliesCount: 3),
-    CommentView(
-        commendId: '1',
-        member: member,
-        createdAt: '2022-2-15T02:54:29.907Z',
-        upVotesCount: 0,
-        downVotesCount: 2,
-        content:
-            'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras quis lacinia est. Morbi elementum pellentesque lectus, eget hendrerit metus lacinia at. Sed sit amet sem blandit, tempor neque pellentesque, ultricies neque. Sed facilisis placerat velit, sit amet cursus lectus posuere luctus. Maecenas vitae efficitur justo. Aenean turpis quam, interdum nec convallis quis, laoreet hendrerit sapien. Nam fringilla ut elit vitae maximus. Donec ac risus in libero lacinia mattis. Maecenas feugiat diam non mauris venenatis consectetur. Nulla ac mi venenatis, sodales lacus ut, elementum nisi.',
-        repliesCount: 2)
-  ];
-}
 
 class ProposalCommentsPage extends StatelessWidget {
   const ProposalCommentsPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final String proposalId = ProposalDetailsBloc().state!.proposalId;
+
     return Scaffold(
       appBar: AppBar(
         title: const Text('Comentarios'),
@@ -62,16 +26,17 @@ class ProposalCommentsPage extends StatelessWidget {
           child: Padding(
             padding: const EdgeInsets.symmetric(vertical: 10),
             child: FutureBuilder(
-              future: _getFakeComments(),
+              future: CommentViewRepository().findByProposalId(proposalId),
               initialData: const <CommentView>[],
               builder: (BuildContext context, AsyncSnapshot snapshot) {
-                final fakeComments = snapshot.data;
+                final comments = snapshot.data;
                 return ListView.builder(
                   physics: const BouncingScrollPhysics(),
-                  itemCount: fakeComments.length,
+                  itemCount: comments.length,
                   itemBuilder: (context, index) => Padding(
                     padding: const EdgeInsets.symmetric(vertical: 8),
-                    child: MemberComment(comment: fakeComments[index]),
+                    child: MemberComment(
+                        comment: comments[index], enableReplies: true),
                   ),
                 );
               },
